Password adding to DB

+++ app/entities/password.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, Text
+from sqlalchemy import BigInteger, Integer, Text
 from sqlalchemy.orm import mapped_column, Mapped
 
 from app.entities.base import Base
@@ -8,6 +8,8 @@ class Password(Base):
     __tablename__ = 'password'
 
     id = mapped_column(Integer, primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     service: Mapped[str] = mapped_column(Text)
     password: Mapped[str] = mapped_column(Text)
+
+
